+++ jest.config.js
@@ -3,7 +3,6 @@ const baseConfig = require("./jest.config.base");
 module.exports = {
     ...baseConfig,
     projects: [
-        "<rootDir>/examples/neo-push/server/jest.config.js",
         "<rootDir>/packages/graphql/jest.config.js",
         "<rootDir>/packages/introspector/jest.config.js",
         "<rootDir>/packages/cypher-builder/jest.config.js",