+++ packages/graphql/src/classes/Node.ts
@@ -17,8 +17,8 @@
  * limitations under the License.
  */
 
-import type { DirectiveNode, NamedTypeNode } from "graphql";
 import camelcase from "camelcase";
+import type { DirectiveNode, NamedTypeNode } from "graphql";
 import pluralize from "pluralize";
 import type {
     ConnectionField,
@@ -36,14 +36,14 @@ import type {
     UnionField,
 } from "../types";
 import type { Auth } from "../types/deprecated/auth/auth";
+import type { DecodedGlobalId } from "../utils/global-ids";
+import { fromGlobalId, toGlobalId } from "../utils/global-ids";
+import { upperFirst } from "../utils/upper-first";
 import type Exclude from "./Exclude";
 import type { GraphElementConstructor } from "./GraphElement";
 import { GraphElement } from "./GraphElement";
 import type { NodeDirective } from "./NodeDirective";
-import type { DecodedGlobalId } from "../utils/global-ids";
-import { fromGlobalId, toGlobalId } from "../utils/global-ids";
 import type { QueryOptionsDirective } from "./QueryOptionsDirective";
-import { upperFirst } from "../utils/upper-first";
 import { NodeAuth } from "./deprecated/NodeAuth";
 
 export interface NodeConstructor extends GraphElementConstructor {
@@ -72,7 +72,6 @@ export interface NodeConstructor extends GraphElementConstructor {
     isGlobalNode?: boolean;
     globalIdField?: string;
     globalIdFieldIsInt?: boolean;
-    federationResolvable: boolean;
 }
 
 type MutableField =
@@ -155,7 +154,6 @@ class Node extends GraphElement {
     public singular: string;
     public plural: string;
     public isGlobalNode: boolean | undefined;
-    public federationResolvable: boolean;
     private _idField: string | undefined;
     private _idFieldIsInt?: boolean;
 
@@ -180,7 +178,6 @@ class Node extends GraphElement {
         this._idFieldIsInt = input.globalIdFieldIsInt;
         this.singular = this.generateSingular();
         this.plural = this.generatePlural(input.plural);
-        this.federationResolvable = input.federationResolvable;
     }
 
     // Fields you can set in a create or update mutation