+++ packages/graphql/tests/tck/issues/3215.test.ts
@@ -17,15 +17,15 @@
  * limitations under the License.
  */
 
-import { gql } from "apollo-server";
+import gql from "graphql-tag";
 import { Neo4jGraphQL } from "../../../src";
-import { formatCypher, translateQuery, formatParams } from "../utils/tck-test-utils";
+import { formatCypher, formatParams, translateQuery } from "../utils/tck-test-utils";
 
 describe("https://github.com/neo4j/graphql/issues/3215", () => {
     describe("union", () => {
         let neoSchema: Neo4jGraphQL;
 
-        const typeDefs = gql`
+        const typeDefs = `#graphql
             type Actor {
                 name: String!
                 age: Int!